pyproject.toml
[project]
name = "account-manager"
version = "0.1.0"
description = "Local account/password manager with AI chatbot interface"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3",
    "langchain-ollama>=0.2",
    "langgraph>=0.2",
    "langchain-community>=0.3",
    "rich>=13.0",
    "cryptography>=42.0",
    "python-frontmatter>=1.1",
    "python-dotenv>=1.0",
    "prompt-toolkit>=3.0",
    "pyyaml>=6.0",
]

[project.scripts]
account-mng = "account_manager.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/account_manager"]
